+++ otelcol/collector_windows.go
@@ -202,7 +202,7 @@ func (w windowsEventLogCore) Sync() error {
 
 func withWindowsCore(elog *eventlog.Log, serviceConfig **service.Config) func(zapcore.Core) zapcore.Core {
 	return func(core zapcore.Core) zapcore.Core {
-		if serviceConfig != nil {
+		if serviceConfig != nil && *serviceConfig != nil {
 			for _, output := range (*serviceConfig).Telemetry.Logs.OutputPaths {
 				if output != "stdout" && output != "stderr" {
 					// A log file was specified in the configuration, so we should not use the Windows Event Log