+++ service/service.go
@@ -34,7 +34,7 @@ import (
 	"go.opentelemetry.io/collector/service/telemetry"
 )
 
-// Settings holds configuration for building a new service.
+// Settings holds configuration for building a new Service.
 type Settings struct {
 	// BuildInfo provides collector start information.
 	BuildInfo component.BuildInfo
@@ -72,6 +72,7 @@ type Service struct {
 	collectorConf     *confmap.Conf
 }
 
+// New creates a new Service, its telemetry, and Components.
 func New(ctx context.Context, set Settings, cfg Config) (*Service, error) {
 	disableHighCard := obsreportconfig.DisableHighCardinalityMetricsfeatureGate.IsEnabled()
 	extendedConfig := obsreportconfig.UseOtelWithSDKConfigurationForInternalTelemetryFeatureGate.IsEnabled()
@@ -88,14 +89,27 @@ func New(ctx context.Context, set Settings, cfg Config) (*Service, error) {
 		},
 		collectorConf: set.CollectorConf,
 	}
-	tel, err := telemetry.New(ctx, telemetry.Settings{BuildInfo: set.BuildInfo, ZapOptions: set.LoggingOptions}, cfg.Telemetry)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get logger: %w", err)
-	}
+
+	// Fetch data for internal telemetry like instance id and sdk version to provide for internal telemetry.
 	res := resource.New(set.BuildInfo, cfg.Telemetry.Resource)
 	pcommonRes := pdataFromSdk(res)
 
-	logger := tel.Logger()
+	telFactory := telemetry.NewFactory()
+	telset := telemetry.Settings{
+		BuildInfo:  set.BuildInfo,
+		ZapOptions: set.LoggingOptions,
+	}
+
+	logger, err := telFactory.CreateLogger(ctx, telset, &cfg.Telemetry)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create logger: %w", err)
+	}
+
+	tracerProvider, err := telFactory.CreateTracerProvider(ctx, telset, &cfg.Telemetry)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create tracer provider: %w", err)
+	}
+
 	logger.Info("Setting up own telemetry...")
 	mp, err := newMeterProvider(
 		meterProviderSettings{
@@ -113,7 +127,7 @@ func New(ctx context.Context, set Settings, cfg Config) (*Service, error) {
 	srv.telemetrySettings = servicetelemetry.TelemetrySettings{
 		Logger:         logger,
 		MeterProvider:  mp,
-		TracerProvider: tel.TracerProvider(),
+		TracerProvider: tracerProvider,
 		MetricsLevel:   cfg.Telemetry.Metrics.Level,
 		// Construct telemetry attributes from build info and config's resource attributes.
 		Resource: pcommonRes,
@@ -247,6 +261,7 @@ func (srv *Service) Shutdown(ctx context.Context) error {
 	return errs
 }
 
+// Creates extensions and then builds the pipeline graph.
 func (srv *Service) initExtensionsAndPipeline(ctx context.Context, set Settings, cfg Config) error {
 	var err error
 	extensionsSettings := extensions.Settings{