+++ service/service_test.go
@@ -320,9 +320,9 @@ func TestServiceTelemetryRestart(t *testing.T) {
 	// check telemetry server to ensure we get a response
 	var resp *http.Response
 
-	// #nosec G107
 	resp, err = http.Get(telemetryURL)
 	assert.NoError(t, err)
+	assert.NoError(t, resp.Body.Close())
 	assert.Equal(t, http.StatusOK, resp.StatusCode)
 	// Response body must be closed now instead of defer as the test
 	// restarts the server on the same port. Leaving response open
@@ -342,8 +342,8 @@ func TestServiceTelemetryRestart(t *testing.T) {
 	// check telemetry server to ensure we get a response
 	require.Eventually(t,
 		func() bool {
-			// #nosec G107
 			resp, err = http.Get(telemetryURL)
+			assert.NoError(t, resp.Body.Close())
 			return err == nil
 		},
 		500*time.Millisecond,