+++ service/telemetry/telemetry_test.go
@@ -55,13 +55,15 @@ func TestTelemetryConfiguration(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			telemetry, err := New(context.Background(), Settings{ZapOptions: []zap.Option{}}, *tt.cfg)
+			f := NewFactory()
+			set := Settings{ZapOptions: []zap.Option{}}
+			logger, err := f.CreateLogger(context.Background(), set, tt.cfg)
 			if tt.success {
 				assert.NoError(t, err)
-				assert.NotNil(t, telemetry)
+				assert.NotNil(t, logger)
 			} else {
 				assert.Error(t, err)
-				assert.Nil(t, telemetry)
+				assert.Nil(t, logger)
 			}
 		})
 	}
@@ -111,10 +113,12 @@ func TestSampledLogger(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			telemetry, err := New(context.Background(), Settings{ZapOptions: []zap.Option{}}, *tt.cfg)
+			f := NewFactory()
+			ctx := context.Background()
+			set := Settings{ZapOptions: []zap.Option{}}
+			logger, err := f.CreateLogger(ctx, set, tt.cfg)
 			assert.NoError(t, err)
-			assert.NotNil(t, telemetry)
-			assert.NotNil(t, telemetry.Logger())
+			assert.NotNil(t, logger)
 		})
 	}
 }